backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
sqlmodel
# paddlepaddle (installed separately)
# paddleocr (installed separately)
pymupdf
groq
python-dotenv
aiofiles
requests
